1.5 !调整了数据库中关于产品和材料关联表的部分! 删除了材料表和原配方表，加入新的配方表，修改逻辑。

+++ src/main/java/com/jason/supplymanagement/service/impl/ProductServiceImpl.java
@@ -1,7 +1,9 @@
 package com.jason.supplymanagement.service.impl;
 
+import com.jason.supplymanagement.dao.ProductComponentDAO;
 import com.jason.supplymanagement.dao.ProductDAO;
 import com.jason.supplymanagement.entity.Product;
+import com.jason.supplymanagement.entity.ProductComponent;
 import com.jason.supplymanagement.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +15,9 @@ public class ProductServiceImpl implements ProductService {
 
     @Autowired
     private ProductDAO productDAO;
+
+    @Autowired
+    private ProductComponentDAO productComponentDAO;
 
     @Override
     public List<Product> getAllProducts() {
@@ -40,6 +45,11 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteProduct(int id) {
+        // Check if the product is used as a component in another product
+        List<ProductComponent> components = productComponentDAO.findByComponent_ProductId(id);
+        if (!components.isEmpty()) {
+            throw new IllegalStateException("Cannot delete product as it is used as a component in another product.");
+        }
         productDAO.deleteById(id);
     }
 }
